chore: improved some code
