+++ apps/web/.env.example
@@ -4,3 +4,4 @@ GOOGLE_CLIENT_ID=""
 GOOGLE_CLIENT_SECRET=""
 NEXTAUTH_URL="http://localhost:3000"
 NEXTAUTH_SECRET="NEXTAUTH_SECRET"
+NEXT_PUBLIC_JUDGE0_API_URL="http://localhost:2358"
