+++ packages/ui/src/code/submission/WrongSubmissionResult.tsx
@@ -1,21 +1,11 @@
-import { Editor } from "@monaco-editor/react";
-import { Clock4, Cpu } from "lucide-react";
+import { useRecoilValue } from "recoil";
+import { activeSubmissionResultSelector } from "@repo/store";
+import SubmissionCreatedAt from "./SubmissionCreatedAt";
+import { Submission, ProblemStatement, TestCase } from "@prisma/client";
 
-const WrongSubmissionResult = ({ submissionResult }: { submissionResult: any }) => {
-  const {
-    statusDesc,
-    createdAt,
-    runtime,
-    memoryUsage,
-    language,
-    code,
-    testCasesPassed,
-    problemStatement,
-    stdout,
-    lastTestCase,
-  } = submissionResult;
-  console.log("SAfcasef", submissionResult);
-
+const WrongSubmissionResult = ({}: {}) => {
+  const submissionDetails = useRecoilValue(activeSubmissionResultSelector);
+  const { statusDesc, createdAt, testCasesPassed, problemStatement, lastTestCase, stdout } = submissionDetails;
   return (
     <div className="p-4 text-[#CCC]">
       <span className="font-bold text-xl text-red-800">{statusDesc}</span>
@@ -24,20 +14,11 @@ const WrongSubmissionResult = ({ submissionResult }: { submissionResult: any }) 
         | &nbsp; &nbsp;
         {testCasesPassed}/{problemStatement.testCases.length} testcases passed
       </span>
-      <div className="text-xs mt-2">
-        submitted at{" "}
-        {new Intl.DateTimeFormat("en-US", {
-          month: "short",
-          day: "numeric",
-          year: "numeric",
-          hour: "2-digit",
-          minute: "2-digit",
-        }).format(new Date(createdAt))}
-      </div>
+      <SubmissionCreatedAt createdAt={createdAt} />
       <div className="my-8 flex flex-col gap-6">
         <div className="flex flex-col gap-3">
           <span className="text-gray-300">Input</span>
-          {lastTestCase.inputs.map((inputParam: any, index: any) => {
+          {lastTestCase.inputs.map((inputParam: string, index: any) => {
             return (
               <div className="flex flex-col gap-1 bg-gray-600  px-3 py-2 rounded-lg" key={index}>
                 <span className="text-gray-300 text-xs">{problemStatement.argumentNames[index]} = </span>
@@ -46,34 +27,19 @@ const WrongSubmissionResult = ({ submissionResult }: { submissionResult: any }) 
             );
           })}
         </div>
+
         <div className="flex flex-col gap-3">
           <span className="text-gray-300">Output</span>
           <div className="flex flex-col gap-1 bg-gray-600  px-3 py-2 rounded-lg">
             <span className="text-sm font-medium">{stdout}</span>
           </div>
         </div>
+
         <div className="flex flex-col gap-3">
           <span className="text-gray-300">Expected</span>
-          <div className="flex flex-col gap-1 border  px-3 py-2 rounded-lg">
+          <div className="flex flex-col gap-1 border bg-gray-600  px-3 py-2 rounded-lg">
             <span className="text-sm font-medium">{lastTestCase.expectedOutput}</span>
           </div>
-        </div>
-      </div>
-      <div className="my-8">
-        <h3>Code | {language.label}</h3>
-        <div className="rounded-md overflow-hidden bg-[#1E1E1E] py-4 my-4 h-[300px]">
-          <Editor
-            height={"100%"}
-            value={atob(code)}
-            theme="vs-dark"
-            options={{
-              fontSize: 14,
-              lineNumbers: "off",
-              readOnly: true,
-              scrollBeyondLastLine: false,
-            }}
-            language={language.value}
-          />
         </div>
       </div>
     </div>
